Removed palette as a texture from the surface & fence renderers in the OXR version, reimplemented as a vec4 array in a new uniform buffer, thus removing the need to use one of the hardware slots to do texture sampling in the device for these particular renderers - it also seemed to improve performance by just a tiny bit in a few initial tests. Upgraded OpenXR Mobile SDK in use to v. 34.0.0 (1.1.50), ktx 1.7.0, and Kotlin 1.6.0 . Downgraded NDK version to 22.1.7171670, since this seems to be the last version that contains an already built Vulkan validation layer library. Cosmetic fix to d_lists to make sure we establish the real identity of the data type behind the lightmap texel data.

+++ id1/d_lists.cpp
@@ -60,7 +60,7 @@ void D_FillSurfaceData (dsurface_t& surface, msurface_t* face, surfcache_t* cach
 	surface.texture_height = texture->height;
 	surface.texture_size = surface.texture_width * surface.texture_height;
 	surface.texture = (unsigned char *)(texture + 1);
-	surface.lightmap_width = cache->width / sizeof(float);
+	surface.lightmap_width = cache->width / sizeof(unsigned);
 	surface.lightmap_height = cache->height;
 	surface.lightmap_size = surface.lightmap_width * surface.lightmap_height;
 	if (d_lists.last_lightmap_texel + surface.lightmap_size >= d_lists.lightmap_texels.size())
@@ -87,7 +87,7 @@ void D_FillSurfaceRotatedData (dsurfacerotated_t& surface, msurface_t* face, sur
 	surface.texture_height = texture->height;
 	surface.texture_size = surface.texture_width * surface.texture_height;
 	surface.texture = (unsigned char *)(texture + 1);
-	surface.lightmap_width = cache->width / sizeof(float);
+	surface.lightmap_width = cache->width / sizeof(unsigned);
 	surface.lightmap_height = cache->height;
 	surface.lightmap_size = surface.lightmap_width * surface.lightmap_height;
 	if (d_lists.last_lightmap_texel + surface.lightmap_size >= d_lists.lightmap_texels.size())
